% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{czasy_wskazniki_pd}
\alias{czasy_wskazniki_pd}
\title{Czasy obliczania wskaznikow w zaleznosci od liczby rdzeni}
\format{
Ramka danych z 51 wierszami i 9 kolumnami:
\itemize{
\item{\code{lWatkowJST} - liczba procesów na poziomie JST (w wywołaniu \code{oblicz_wskazniki_pd_jst()}),}
\item{\code{lWatkowGrupy} - liczba procesów na poziomie grup (w wywołaniu \code{oblicz_wskazniki_grupy()} - w ramach każdego procesu na poziomie JST).}
\item{\code{lWatkowOgolem} - łączna liczba równoległych procesów.}
\item{\code{poziom} - poziom agregacji JST (w wywołaniu \code{oblicz_wskazniki_pd_jst()}).}
\item{\code{czas} - czas obliczania wskaźników [s].}
\item{\code{czasRel1.1} - iloraz czasu obliczania wskaźników do czasu obliczania wskaźników na tym samym poziomie agregacji JST, bez zrównoleglania.}
\item{\code{czasRelJST.1} - iloraz czasu obliczania wskaźników do czasu obliczania wskaźników na tym samym poziomie agregacji JST, z taką samą liczbą równoległych wątków na poziomie JST i bez zrównoleglania na poziomie grup.}
\item{\code{czasRel1.Grupy} - iloraz czasu obliczania wskaźników do czasu obliczania wskaźników na tym samym poziomie agregacji JST, bez zrównoleglania na poziomie JST i z taką samą liczbą równoległych wątków na poziomie grup.}
\item{\code{czasRel1Ogolem} - iloraz czasu obliczania wskaźników do czasu obliczania wskaźników na tym samym poziomie agregacji JST, z tą samą liczbą równoległych wątków ogółem i bez zrównoleglania na poziomie JST.}
}
}
\usage{
czasy_wskazniki_pd
}
\description{
Zestawienie czasów obliczania wskaźników przez
\code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}} w zależności od wartości argumentów
\code{liczbaWatkow} i \code{liczbaWatkowGrupy}. Wykorzystano dane dla rocznika
absolwentów 2020 z edycji monitoringu 2022. Jako zestawu zmiennych
grupujących użyto: \code{typ_szk}, \code{szk_specjalna}, \code{kod_zaw}, \code{nazwa_zaw},
\code{plec}, \code{mlodoc}, przy domyślnym zestawie obliczanych wskaźników. Obliczenia
zostały wykonane na laptopie z dwunastordzeniowym procesorem AMD Ryzen 5 5600H
z 64 GB RAM.

Dla każdej kombinacji wartości parametrów obliczenia uruchamiano tylko raz,
nie dbając przesadnie o standaryzację warunków - niekiedy (gdy
wykorzystywanych było łącznie mniej rdzeni) w tle działała również
przeglądarka internetowa i aplikacja poczty elektornicznej, więc \strong{wyniki nie
są zbyt dokładne}.
}
\seealso{
\code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}}
}
\keyword{datasets}
