% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodaj_procenty_wskazniki_pd.R
\name{dodaj_procenty_wskazniki_pd}
\alias{dodaj_procenty_wskazniki_pd}
\title{Przeksztalcanie wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
dodaj_procenty_wskazniki_pd(x, dopiszZadenZWymienionych = FALSE)
}
\arguments{
\item{x}{ramka danych zawierająca (między innymi) kolumny ze zagregowanymi
wskaźnikami przeznaczonymi do publicznej prezentacji, typowo zwrócona przez
funkcje \code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}} lub
\code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}}}

\item{dopiszZadenZWymienionych}{wartość logiczna - czy jeśli wskaźnik
posiada atrybut \code{lZadenZWymienionych}, to powinien on zostać dodany do
rozkładu (z nazwą "Żaden z wymienionych")}
}
\value{
Ramka danych przekazana argumentem \code{x}, w której wskaźniki będące
rozkładami (a przynajmniej na takie wyglądające) zostały uzupełnione
o częstości. Co do zasady przekształcony wskaźnik ma dotychczasowe nazwy
w wierszach, a w kolumnach rodzaj rozkładu.
}
\description{
Funkcja pozwala zmienić przekształcić wskaźniki zagregowane będące rozkładami
liczebności poprzez dopisanie do nich rozkładu częstości.
}
\details{
Funkcja przekształca \strong{wszystko, co \emph{wyglądają jej na zagregowany wskaźnik
przeznaczony do publicznej prezentacji będący rozkładem liczebności}}, co
w praktyce oznacza, że sprawdzane są elementy kolumn-list w ramce danych
przekazanej argumentem \code{x}. Jeśli element takiej kolumny-listy jest tablicą
(\emph{table}) lub ma atrybut \code{lZadenZWymienionych}, to zostanie uznany za
kwalifikujący się do przekształcenia. Zostanie on przekształcony w tablicę
(\emph{array}) z dodanym dodatkowym wymiarem o nazwie "rozklad", którego dwa
elementy będą miały nazwy "N" i "Pct", gdzie element "N" będzie przechowywał
dotychczasową postać wskażnika, a element "Pct" dotychczasowe wartości
podzielone przez wartość atrybutu \code{lAbs}. Atrybuty \code{lAbs}, \code{lSzk},
\code{lZadenZWymienionych} i \code{lNieDotyczy} zostaną w przekształconym wskażniku
bez zmian.

Jeśli argument \code{dopiszZadenZWymienionych=TRUE}, przed dokonaniem opisanego
wyżej przekształcenia, wartość atrybutu \code{lZadenZWymienionych}
(o ile istnieje) zostanie dopisana jako dodatkowy element na końcu wskaźnika.
}
