% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodaj_wartosc_ogolem.R
\name{dodaj_wartosc_ogolem}
\alias{dodaj_wartosc_ogolem}
\title{Funkcje pomocnicze}
\usage{
dodaj_wartosc_ogolem(x, wartosc, zawszeDodawaj = TRUE)
}
\arguments{
\item{x}{wektor lub czynnik}

\item{wartosc}{wartość, która ma zostać podmieniona w miejsce braków danych
(jednoelementowy wektor)}

\item{zawszeDodawaj}{opcjonalnie wartość logiczna - czy \code{wartosc} powinna
zostać dodana również wtedy, kiedy \code{x} ma tylko jedną unikalną wartość lub
jest czynnikiem o tylko jednym poziomie; domyślnie \code{TRUE}}
}
\value{
\strong{Funkcja zawsze zwraca czynnik}, którego zestaw poziomów oprócz
wcześniej istniejących, lub - jeśli argumentem \code{x} został przekazany wektor -
zbiór wartości występujących w danych, został rozszerzony o wartość podaną
argumentem \code{wartosc} (chyba że \code{zawszeDodawaj=FALSE}).
}
\description{
Pozwala dodać do zestawu wartości danej zmiennej (wektora lub czynnika)
\emph{wartość ogółem}.
}
