% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodaj_wskazniki_dyplomy.R
\name{dodaj_wskazniki_dyplomy}
\alias{dodaj_wskazniki_dyplomy}
\title{Obliczanie wskaznikow wykorzystywanych w raportach}
\usage{
dodaj_wskazniki_dyplomy(p4, p1, maksMiesOdUkoncz, rokMonitoringu = NULL)
}
\arguments{
\item{p4}{ramka danych z tabelą \emph{pośrednią} P4 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs}
i \code{kod_zaw}}

\item{p1}{ramka danych z tabelą \emph{pośrednią} P1 (lub jej odfiltrowanym
podzbiorem)}

\item{maksMiesOdUkoncz}{liczba całkowita opisująca okres, który będzie brany
pod uwagę: \strong{jeśli jest to liczba nieujemna}, będzie to maksymalna wartość
zmiennej \code{mies_od_ukoncz} w ramce danych przekazanej argumentem \code{p1};
\strong{jeśli jest to liczba ujemna}, konieczne jest również podanie wartości
argumentu \code{rokMonitoringu} a brany pod uwagę okres zostanie ograniczony
z góry do czerwca \code{rokMonitoringu} plus (ujemna) \code{maksMiesOdUkoncz} (czyli
np. wartość -3 wskazuje na maksymalnie marzec roku prowadzenia monitoringu)}

\item{rokMonitoringu}{opcjonalnie liczba całkowita - rok (edycja)
monitoringu; musi być podana, jeśli \code{maksMiesOdUkoncz} jest liczbą ujemną
(w przeciwnym wypadku wartość tego argumentu nie jest w żaden sposób
wykorzystywana)}
}
\value{
Ramka danych przekazana argumentem \code{p4} z dodanymi kolumnami:
\itemize{
\item \code{matura_zdana} - liczba całkowita: 0 - nie, 1 - tak,
\item \code{dyplom_zaw} - czynnik o poziomach: "Świadectwo czeladnicze",
"Dyplom zawodowy", "Tylko certyfikat kwalifikacji",
"Brak certyfikatów i dyplomu".
}
}
\description{
Na podstawie tabeli \emph{pośredniej} P1 funkcja oblicza wskaźniki opisujące
uzyskiwanie dyplomów: zdanie matury oraz uzyskanie dyplomów zawodowych
(świadectw czeladniczych, certyfikatów kwalifikacji).
}
\details{
W przypadku \code{dyplom_zaw}:
\itemize{
\item Pod uwagę brane jest tylko uzyskanie dyplomu/certyfikatu/świadectwa
w wyuczonym zawodzie;
\item W przypadku dyplomów zawodowych dodatkowo brane są pod uwagę tylko te
uzyskane nie wcześniej niż w roku ukończenia szkoły;
\item Uzyskanie dyplomu zawodowego nadpisuje wszystkie inne
certyfikaty/świadectwa, a uzyskanie tytułu czeladnika nadpisuje
uzyskanie certyfikatu kwalifikacji.
\item Osoby, które nie kształciły się zawodowo mają przypisany brak danych.
}

W przypadku \code{matura_zdana} wartości obliczane są dla wszystkich rekordów
\strong{bez względu na typ ukończonej szkoły}.
}
