% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodaj_wskazniki_kontynuacje.R
\name{dodaj_wskazniki_kontynuacje}
\alias{dodaj_wskazniki_kontynuacje}
\title{Obliczanie wskaznikow wykorzystywanych w raportach}
\usage{
dodaj_wskazniki_kontynuacje(
  p4,
  p2,
  miesOdUkoncz,
  tylkoLegalne = TRUE,
  wyroznijTrybStudiow = FALSE
)
}
\arguments{
\item{p4}{ramka danych z tabelą \emph{pośrednią} P4 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs}
i \code{typ_szk}}

\item{p2}{ramka danych z tabelą \emph{pośrednią} P2 (lub jej odfiltrowanym
podzbiorem)}

\item{miesOdUkoncz}{wektor liczb całkowitych - wartości zmiennej
\code{mies_od_ukoncz}, dla których mają zostać obliczone wskaźniki}

\item{tylkoLegalne}{wartość logiczna - czy usunąć z danych przekazanych \code{p2}
kontynuacje, które nominalnie powinny być niemożliwe (ze względu na typ
szkoły ukończonej przez absolwenta, por. \link{legalneKontynuacje})?}

\item{wyroznijTrybStudiow}{wartość logiczna - czy w konstruowanych
wskaźnikach wyróżniać studia stacjonarne i niestacjonarne? (oczywiście, o ile
takie rozróżnienie jest dostępne w przekazanych danych, tj. od edycji
monitoringu 2026)}
}
\value{
Ramka danych przekazana argumentem \code{p4} z dodanymi kolumnami:
\itemize{
\item \verb{typ_szk_kont<mies_od_ukoncz>} - formy kontynuacji nauki,
\item \verb{branza_kont_bsii<mies_od_ukoncz>} - branża kontynuacji nauki w BS II przez absolwentów BS I,
\item \verb{dziedzina_kont<mies_od_ukoncz>} - dziedzina kontynuacji nauki na studiach,
\item \verb{dyscyplina_kont<mies_od_ukoncz>} - dyscyplina kontynuacji nauki na studiach.
}
}
\description{
Na podstawie tabeli \emph{pośredniej} P2 funkcja oblicza wskaźniki opisujące
formy kontynuowania nauki.
}
\details{
Każda z tworzonych kolumn jest macierzą binarną, w której wartość 1 opisuje
kontynuowanie nauki w danej formie/branży/dziedzinie/dyscyplinie,
0 niekontynuowanie, a brak danych, że dana kolumna nie ma sensu dla danego
absolwenta ze względu na typ ukończonej przez niego szkoły (choć jeśli
argument \code{tylkoLegalne} został ustawiony na \code{FALSE}, to podjęcie nauki
w danej formie itp. zostanie odnotowane wartością 1, nawet jeśli nominalnie
taka sytuacja nie powinna się zdarzyć).
}
\examples{
\dontrun{
  p4 <- dodaj_wskazniki_kontynuacje(p4, p2, c(6L, 18L, 30L, 42L, 54L))
  colSums(p4$typ_szk_kont6, na.rm = TRUE)
  colSums(dodaj_wskazniki_kontynuacje(p4, p2, 6L, tylkoLegalne = FALSE)$typ_szk_kont6,
          na.rm = TRUE)
}
}
\seealso{
\code{\link[=oblicz_wskaznik_macierz]{oblicz_wskaznik_macierz()}}- \emph{koń roboczy} odpowiedzialny za
obliczanie wskaźników w ramach \code{dodaj_wskazniki_kontynuacje()}
}
