% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopisz_wskaznik_pd_liczba_abs.R
\name{dopisz_wskaznik_pd_liczba_abs}
\alias{dopisz_wskaznik_pd_liczba_abs}
\title{Agregowanie wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
dopisz_wskaznik_pd_liczba_abs(x, uzyjWskaznik, uzyjCzas = NA)
}
\arguments{
\item{x}{ramka danych z wartościami wskaźników zwrócona przez
\code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}} lub
\code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}}}

\item{uzyjWskaznik}{ciąg znaków z nazwą wskaźnika, na podstawie
którego ma zostać utworzony wskaźnik opisujący liczbę absolwentów}

\item{uzyjCzas}{opcjonalnie wektor zawierający wartości zmiennej \code{czas}
(w danych przekazanych argumentem \code{x}), dla których ma zostać obliczony nowy
wskaźnik; domyślna wartość - brak danych - wskazuje, że mają zostać użyte
wszystkie wiersze (opisujące wskaźnik, na który wskazuje argument
\code{uzyjWskaznik}) występujące w przekazanych danych}
}
\value{
Ramka danych przekazana argumentem \code{x} z dodanymi wierszami,
w których zmienna \code{wskaźnik} przyjmuje wartość "liczba_abs", zmienna
\code{czas} przyjmuje wartość "0", a zmienna \code{wartosc} jest listą
jednoelementowych wektorów liczb (każdy taki pojedynczy element ma nazwę
"liczba") i atrybutami \code{lAbs} i \code{lSzk}.
}
\description{
Funkcja pozwala dodać do ramki danych z wartościami wskaźników zwróconej
przez \code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}} lub
\code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}} dodatkowe wiersze ze wskaźnikiem "liczba_abs",
opisującym liczbę absolwentów, a utworzonym na podstawie informacji o liczbie
absolwentów zawartej w atrybutach jednego ze wskaźników już występujących
w przekazanych danych.
}
\details{
Nowe wiersze opisują wszystkie występujące w ramce danych przekazanej
argumentem \code{x} kombinacje wartości \emph{zmiennych grupujących} (a więc wszystkich
poza \code{wskaznik}, \code{czas} i \code{wartosc}), dla których kolumna \code{wskaźnik}
przyjmuje wartość podaną argumentem \code{uzyjWskaznik}, a wartość kolumny \code{czas}
należy do zbioru wartości podanego argumentem \code{uzyjCzas}.
}
\seealso{
\code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}},
\code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}}, \code{\link[=zanonimizuj_wskazniki_pd]{zanonimizuj_wskazniki_pd()}}
}
