% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_wskaznik_macierz.R
\name{oblicz_wskaznik_macierz}
\alias{oblicz_wskaznik_macierz}
\title{Obliczanie wskaznikow opisujacych (potencjalnie) wspolwystepujace stany}
\usage{
oblicz_wskaznik_macierz(
  x,
  zm,
  zestawWartosci = unique(c(levels(x[[zm]]), sort(unique(x[[zm]])))),
  wszystkieObs = NULL
)
}
\arguments{
\item{x}{ramka danych z tabelą \emph{pośrednią} (lub jej odfiltrowanym
podzbiorem); musi zawierać kolumny \code{id_abs} i \code{rok_abs} oraz kolumnę wskazaną
wartością argumentu \code{zm}}

\item{zm}{ciąg znaków - nazwa zmiennej w \code{x}, na podstawie której ma zostać
obliczony wskaźnik}

\item{zestawWartosci}{opcjonalnie ciąg znaków zawierający zestaw wartości,
dla których ma zostać obliczony wskaźnik; domyślnie jest to albo zbiór
poziomów zmiennej o nazwie \code{zm} - jeśli jest ona czynnikiem - albo
posortowany zbiór jej unikalnych wartości w danych przekazanych argumentem
\code{x} - jeśli nie jest czynnikiem}

\item{wszystkieObs}{opcjonalnie ramka danych zawierająca wszystkie kombinacje
wartości zmiennych identyfikujących \emph{obserwacje}, dla których mają zostać
obliczone wartości wskaźnika, nawet jeśli niektóre z nich nie występują
w danych przekazanych argumentem \code{x}; do identyfikacji takich kombinacji
zostaną wykorzystane wszystkie zmienne występujące zarówno w obiekcie
przekazanym argumentem \code{x}, jak i w obiekcie przekazanym argumentem
\code{wszystkieObs} (z wyjątkiem zmiennej o nazwie podanej argumentem \code{zm})}
}
\value{
Ramka danych o kolumnach takich jak w \code{x}, lub - jeśli podano
argument \code{wszystkieObs} - kolumnach występujących zarówno w \code{x} jak
i we \code{wszystkieObs} oraz kolumnie o nazwie podanej argumentem \code{zm}, z tym
że kolumna \code{zm} jest macierzą, której kolumny mają nazwy podane argumentem
\code{zestawWartosci}.  Liczba wierszy odpowiada albo liczbie wierszy obiektu
przekazanego argumentem \code{wszystkieObs} albo liczbie unikalnych kombinacji
wartości zmiennych nie będących zmienną o nazwie przekazanej argumentem \code{zm}
w obiekcie przekazanym argumentem \code{x}.
}
\description{
Na podstawie tabeli \emph{pośredniej} przekazanej argumentem \code{x} oblicza wartości
wskaźnika na podstawie zmiennej (kolumny) wskazanej argumentem \code{zm} dla
poszczególnych kombinacji wartości pozostałych zmiennych występujących
w przekazanej argumentem \code{x} ramce danych lub - jeśli został przekazany
argument \code{wszystkieObs} - na podstawie kombinacji wartości zmiennych
wspólnych dla obiektów przekazanych argumentami \code{x} i \code{wszystkieObs}.
}
\details{
Funkcja służy do obliczania wskaźników, które opisują kilka stanów mogących
współwystępować jednocześnie, np. kontynuacja nauki w różnych formach,
czy studiowanie w poszczególnych dziedzinach nauki. Obliczany wskaźnik ma
postać macierzy 0 i 1, której kolumny mają nazwy podane argumentem
\code{zestawWartosci}, w której wartość 1 wskazuje na zaistnienie danego zdarzenia
a wartość 0 na niezaistnienie.
}
\seealso{
\code{\link[=dodaj_wskazniki_kontynuacje]{dodaj_wskazniki_kontynuacje()}}
}
