% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_wskaznik_z_p3.R
\name{oblicz_wskaznik_z_p3}
\alias{oblicz_wskaznik_z_p3}
\title{Agregowanie wskaznikow po miesiacach w ramach absolwentow}
\usage{
oblicz_wskaznik_z_p3(
  p3,
  zm,
  wszystkieObs = NULL,
  fun = mean,
  pomijaj0 = grepl("wynagrodzenie", zm)
)
}
\arguments{
\item{p3}{ramka danych z tabelą \emph{pośrednią} P3 (lub jej odfiltrowanym
podzbiorem); musi zawierać kolumny \code{id_abs} i \code{rok_abs} oraz kolumnę wskazaną
wartością argumentu \code{zm}}

\item{zm}{ciąg znaków - nazwa zmiennej w \code{x}, która ma podlegać agregacji
po czasie w ramach absolwentów (obserwacji)}

\item{wszystkieObs}{opcjonalnie ramka danych zawierająca wszystkie kombinacje
wartości zmiennych identyfikujących \emph{obserwacje}, dla których mają zostać
obliczone wartości wskaźnika, nawet jeśli niektóre z nich nie występują
w danych przekazanych argumentem \code{p3}; do identyfikacji takich kombinacji
zostaną wykorzystane wszystkie zmienne występujące zarówno w obiekcie
przekazanym argumentem \code{p3}, jak i w obiekcie przekazanym argumentem
\code{wszystkieObs} (z wyjątkiem zmiennej o nazwie podanej argumentem \code{zm})}

\item{fun}{funkcja lub jednostronna formuła z wyrażeniem, które ma zostać
użyte do agregacji wartości zmiennej o nazwie \code{zm}; ponieważ braki danych są
przez \code{oblicz_wskaznik_z_p3()} usuwane z danych przed przekazaniem do
agregacji, nie jest więc konieczne zadawanie ich obsługi w ramach tego
argumentu}

\item{pomijaj0}{opcjonalnie wartość logiczna - czy z agregacji powinny zostać
wyłączone wartości zmiennej o nazwie \code{zm} równe 0? domyślnie \code{TRUE} przy
agregacji wynagrodzeń i \code{FALSE} w pozostałych przypadkach}
}
\value{
Ramka danych o kolumnach takich jak w \code{p3}, lub - jeśli podano
argument \code{wszystkieObs} - kolumnach występujących zarówno w \code{p3} jak
i we \code{wszystkieObs} oraz kolumnie o nazwie podanej argumentem \code{zm}. Liczba
wierszy odpowiada albo liczbie wierszy obiektu przekazanego argumentem
\code{wszystkieObs} albo liczbie unikalnych kombinacji wartości zmiennych
(\code{id_abs}, \code{rok_abs}) w obiekcie przekazanym argumentem \code{p3}.
}
\description{
Na podstawie tabeli \emph{pośredniej} przekazanej argumentem \code{x} oblicza wartości
wskaźnika na podstawie zmiennej (kolumny) wskazanej argumentem \code{zm} dla
poszczególnych kombinacji wartości pozostałych zmiennych występujących
w przekazanej argumentem \code{x} ramce danych lub - jeśli został przekazany
argument \code{wszystkieObs} - na podstawie kombinacji wartości zmiennych
wspólnych dla obiektów przekazanych argumentami \code{x} i \code{wszystkieObs}.
}
\details{
Funkcja służy do agregowania wskaźników po czasie w ramach absolwentów,
np. wskaźnika wynagrodzeń (przychodów). Funkcja zakłada, że dane przekazane
argumentem \code{p3} zostały już zawężone do odpowiedniego okresu i że różne
wiersze w ramach danej kombinacji wartości zmiennych (\code{id_abs}, \code{rok_abs})
reprezentują poszczególne miesiące.
}
\seealso{
\code{\link[=dodaj_wskazniki_prace]{dodaj_wskazniki_prace()}}
}
