% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_wskazniki_pd.R
\name{oblicz_wskazniki_pd}
\alias{oblicz_wskazniki_pd}
\title{Agregowanie wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
oblicz_wskazniki_pd(
  p4,
  p3,
  zmWskaznikiP4 = c("dyplom_zaw", "matura_zdana", "typ_szk_kont6", "typ_szk_kont18",
    "dyscyplina_kont6", "dyscyplina_kont18", "sr_wynagr_r1"),
  zmWskaznikiP3 = c("status", "bezrobocie"),
  statystyki = list(średnia = ~mean(., na.rm = TRUE)),
  wskTylkoNiezerowe = c("dziedzina_kont", "dyscyplina_kont"),
  oddzielCzasOdWskaznika = TRUE,
  format = c("długi", "szeroki"),
  miesiace = NULL
)
}
\arguments{
\item{p4}{ramka danych z tabelą \emph{pośrednią} P4 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs} i \code{id_szk}}

\item{p3}{ramka danych z tabelą \emph{pośrednią} P3 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs}
i \code{mies_od_ukoncz} lub \code{NULL} (jeśli \code{zmWskaznikiP3 } nie podaje żadnych
wskzaźników)}

\item{zmWskaznikiP4}{opcjonalnie wektor ciągów znaków z nazwami zmiennych,
które zawierają wskaźniki w danych przekazanych argumentem \code{p4}, które mają
zostać zagregowane; jeśli argument nie zostanie podany, zostanie wykorzystany
domyślny zestaw zmiennych; można podać \code{NULL} lub wektor tekstowy zerowej
długości, aby wskazać, że żadne wskaźniki z \code{p4} nie mają być agregowane}

\item{zmWskaznikiP3}{opcjonalnie wektor ciągów znaków z nazwami zmiennych,
które zawierają wskaźniki w danych przekazanych argumentem \code{p3}, które mają
zostać zagregowane; jeśli argument nie zostanie podany, zostanie wykorzystany
domyślny zestaw zmiennych; można podać \code{NULL} lub wektor tekstowy zerowej
długości, aby wskazać, że żadne wskaźniki z \code{p3} nie mają być agregowane}

\item{statystyki}{opcjonalnie lista z nazwanymi elementami, opisującymi
statystyki, które mają zostać obliczone w odniesieniu do wskaźników
\strong{liczbowych} (w formacie takim, jak argument \code{.fns} w formie listy
w wywołaniach \code{\link[dplyr:across]{dplyr::across()}}); funkcje/wyrażenia mogą zwracać więcej niż
jedną wartość, ale ew. nazwy elementów zwróconego przez funkcję wektora nie
zostaną zachowane; domyślnie średnia obliczana z pominięciem
braków danych}

\item{wskTylkoNiezerowe}{opcjonalnie wektor ciągów znaków z nazwami
\strong{wskaźników} obliczanych w formie rozkładów (a więc na podstawie zmiennych
będących czynnikami, wektorami tekstowymi lub macierzami), dla których mają
zostać zwrócone tylko wartości rozkładu nie będące zerami ani brakami danych}

\item{oddzielCzasOdWskaznika}{opcjonalnie wartość logiczna - czy w zwracanej
ramce danych nazwa wskaźnika powinna zostać rozbita na dwie kolumny:
\code{wskaznik} i \code{czas} (\code{TRUE}), czy też tylko kolumnę \code{wskaznik}, zawierającą
nazwę wskaźnika połączoną z kodem okresu, dla jakiego został on obliczony
(\code{FALSE});

jeśli nazwa agregowanej zmiennej nie zawiera przyrostka wskazującego na czas,
do jakiego się odnosi, to obliczone na jej podstawie wskaźniki w zwróconych
wynikach albo otrzymają przyrostek "0" (jeśli \code{oddzielCzasOdWskaznika=FALSE}
lub \code{format="szeroki"}) albo będą miały przypisaną wartość "0" w kolumnie \code{czas}}

\item{format}{opcjonalnie ciąg znaków opisujący, czy wyniki mają zostać
zwrócone w formie \emph{długiej} (domyślnie), czy \emph{szerokiej}}

\item{miesiace}{opcjonalnie wektor liczb całkowitych (co do zasady dodatnich)
opisujący, wiersze o jakich wartościach kolumny \code{mies_od_ukoncz} z ramki
danych przekazanej argumentem \code{p3} mają zostać wykorzystane do obliczenia
zagregowanych wskaźników (wszystkie inne zostaną pominięte); pozwala też
uzyskać efekt polegający na utworzeniu pożądanego zestawy wskaźników
obliczonych na podstawie danych z tabeli \emph{pośredniej} P3 w sytuacji, gdy
dane przekazane do funkcji są puste (mają zero wierszy); domyślnie zostaną
wykorzystane wszystkie dane z \code{p3}}
}
\value{
Ramka danych z obliczonymi wskaźnikami, w zależności od kombinacji
wartości argumentów \code{oddzielCzasOdWskaznika} i \code{format} zawierająca:
\itemize{
\item wiele wierszy i trzy kolumny: \code{wskaznik}, \code{czas} i \code{wartosc}
(domyślnie: \code{oddzielCzasOdWskaznika=TRUE} i \code{format="długi"}),
\item wiele wierszy i dwie kolumny: \code{wskaznik} i \code{wartosc}
(\code{oddzielCzasOdWskaznika=FALSE} i \code{format="długi"}),
\item wiele wierszy, kolumnę \code{czas} i kolejne kolumny zawierające poszczególne
wskaźniki (\code{oddzielCzasOdWskaznika=TRUE} i \code{format="szeroki"}),
\item jeden wiersz i kolejne kolumny zawierające poszczególne
wskaźniki (\code{oddzielCzasOdWskaznika=FALSE} i \code{format="szeroki"}).
}

Odnośnie do formatu kolumn zawierających wartości wskaźników, patrz niżej.
}
\description{
Funkcja pozwala obliczyć zestawienia zagregowanych wskaźników z danej edycji
monitoringu na podstawie wskazanych zmiennych w przekazanych danych.
}
\details{
Agregacja przeprowadzana jest w różny sposób w zależności od typu kolumny
w wejściowych danych:
\itemize{
\item Kolumny będące czynnikami lub wektorami tekstowymi (w sensie sprawdzania
typu z wykorzystaniem funkcji \code{\link[=is.factor]{is.factor()}} i \code{\link[=is.character]{is.character()}}) zostaną
zagregowane z wykorzystaniem funkcji \code{\link[=table]{table()}}.
\itemize{
\item Przy pomocy argumentu \code{wskTylkoNiezerowe} można wskazać, dla których
wskaźników mają zostać zwrócone tylko te wartości rozkładu, które
przyjmują niezerowe (i niebędące brakami danych) liczebności.
}
\item Kolumny będące wektorami liczbowymi lub logicznymi (w sensie sprawdzania
typu z wykorzystaniem funkcji \code{\link[=is.numeric]{is.numeric()}} i \code{\link[=is.logical]{is.logical()}}) zostaną
zagregowane przy pomocy funkcji/wyrażeń przekazanych argumentem
\code{statystyki} - efektem będzie wektor, nazwy elementów którego
odpowiadają nazwom elementów listy przekazanej argumentem \code{statystyki}
(domyślnie jednoelementowy wektor, z nazwą elementu "średnia").
\item Kolumny będące macierzami liczb lub wartości logicznych zostaną
zagregowane poprzez zsumowanie wartości w kolumnach, przy czym
\strong{kolumny, w których znajdowały się tylko braki danych, zostaną usunięte
z wyników} (nie dotyczy to sytuacji, gdy dane są puste, tzn. mają zero
wierszy, bo wtedy nie ma sensu rozważanie, czy w danej kolumnie znajduje
się brak danych, czy nie).
\itemize{
\item Przy pomocy argumentu \code{wskTylkoNiezerowe} można wskazać, dla których
wskaźników mają zostać zwrócone tylko te wartości rozkładu, które
przyjmują niezerowe (i niebędące brakami danych) liczebności.
}
}

W efekcie każda kolumna zawierająca zagregowane wskaźniki w zwracanej ramce
danych jest listą. Każdy z jej elementów stanowi wskaźnik obliczony zgodnie
z powyższym opisem, który ma dodatkowo przypisane atrybuty
(p. \code{\link[=attributes]{attributes()}}), z których każdy jest liczbą naturalną:
\itemize{
\item \code{lAbs} - liczba wierszy (absolwentów), na podstawie których został
obliczony zagregowany wskaźnik,
\item \code{lSzk} - liczba różnych szkół, które ukończyli absolwenci, na podstawie
których został obliczony zagregowany wskaźnik,
\item \code{lNieDotyczy} - liczba wierszy (absolwentów), zawierających braki danych
(w przypadku kolumn wejściowych będących macierzami - we wszystkich
kolumnach macierzy),
\item Wskaźniki zagregowane obliczone na podstawie kolumn macierzowych
mają dodatkowy atrybut \code{lZadenZWymienionych}, który przechowuje
liczbę wierszy, w których wszystkie kolumny macierzy były albo
zerami albo brakami danych (ale z pominięciem wierszy zawierających
same braki danych).
}
}
\seealso{
\code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}},
\code{\link[=dopisz_wskaznik_pd_liczba_abs]{dopisz_wskaznik_pd_liczba_abs()}}, \code{\link[=zanonimizuj_wskazniki_pd]{zanonimizuj_wskazniki_pd()}}
}
