% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podmien_braki_danych.R
\name{podmien_braki_danych}
\alias{podmien_braki_danych}
\title{Funkcje pomocnicze}
\usage{
podmien_braki_danych(x, wartosc, zawszeDodajPoziom = TRUE)
}
\arguments{
\item{x}{wektor lub czynnik}

\item{wartosc}{wartość, która ma zostać podmieniona w miejsce braków danych
(jednoelementowy wektor)}

\item{zawszeDodajPoziom}{opcjonalnie wartość logiczna - czy jeśli \code{x} jest
czynnikiem, to dodawać do zestawu jego poziomów \code{wartość} nawet wtedy, jeśli
w \code{x}, ani wśród jego poziomów, nie występują żadne braki danych? jeśli \code{x}
nie jest czynnikiem, argument jest ignorowany; domyślnie \code{TRUE}}
}
\value{
Obiekt przekazany argumentem \code{x} z podmienionymi brakami danych
(a w przypadku czynników również odpowiednio zmienionym zestawem poziomów
czynnika).
}
\description{
Pozwala podmienić braki danych na podaną wartość. W odróżnieniu od wielu
innych podobnych funkcji \strong{obsługuje również czynniki}.
}
