% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_dane_przeplywy.R
\name{przygotuj_dane_przeplywy}
\alias{przygotuj_dane_przeplywy}
\title{Przygotowanie danych do tworzenia wykresu przeplywow}
\usage{
przygotuj_dane_przeplywy(p3, zm, punktyCzasu, zmPunktyCzasu = "mies_od_ukoncz")
}
\arguments{
\item{p3}{ramka danych z tabelą \emph{pośrednią} P3 (lub jej odfiltrowanym
podzbiorem)}

\item{zm}{<\link[dplyr:dplyr_tidy_select]{tidy-select}> nazwa kolumny opisującej
statusy (pomiędzy którymi może dochodzić do przepływów)}

\item{punktyCzasu}{wektor wartości, dla których mają być rozpatrywane statusy
\itemize{
\item domyślnie wartości zmiennej \code{mies_od_ukoncz}
}}

\item{zmPunktyCzasu}{<\link[dplyr:dplyr_tidy_select]{tidy-select}> opcjonalnie
nazwa zmiennej zawierającej wartości przekazane argumentem \code{punktyCzasu}
(domyślnie \code{mies_od_ukoncz})}
}
\value{
Ramka danych o kolumnach:
\itemize{
\item \code{idKombinacji} - identyfikator kombinacji (do przypisania \emph{estetyce} \code{alluvium} w wywołaniach funkcji pakietu \emph{ggalluvial}),
\item \verb{<zmPunktyCzasu>} - zmienna wskazana argumentem \code{zmPunktyCzasu} (do przypisania \emph{estetyce} \code{x} albo \code{y} w wywołaniach funkcji pakietu \emph{ggalluvial}),
\item \verb{<zm>} - zmienna wskazana argumentem \code{zm} (do przypisania \emph{estetyce} \code{stratum} w wywołaniach funkcji pakietu \emph{ggalluvial}),
\item \code{n} - liczba absolwentów o trajektorii opisywanej \strong{daną wartością \code{idKombinacji}},
\item \code{pct} - procent absolwentów \strong{w danym punkcie czasu} o statusie opisanym wartością zmiennej \verb{<zm>} (do przypisania \emph{estetyce} \code{y} albo \code{x} w wywołaniach funkcji pakietu \emph{ggalluvial}),
\item \code{pctWarstwa} - procent absolwentów o trajektorii opisywanej daną wartością \code{idKombinacji} \strong{wśród absolwentów o danym statusie w danym punkcie czasu} (do przypisania \emph{estetyce} \code{weight} w wywołaniach funkcji \code{stat_flow()} pakietu \emph{ggalluvial} mających na celu pokazanie etykiet wartości poszczególnych przepływów)
}
}
\description{
Funkcja pozwala zagregować dane - domyślnie pochodzące z tabeli \emph{pośredniej}
P3 - do postaci, w której mogą one zostac łatwo wykorzystane do przygotowania
wykresu przepływów, w szczególności z wykorzystaniem pakietu \emph{ggalluvial}
(i korzystającego z tego pakietu szablonu wykresu \code{wykresPrzeplywyStatusy},
zawartego w pakiecie \emph{LOSYkolory}).
}
\details{
Funkcja może być wykorzystana do przekształcenia również innych danych, niż
tabela \emph{pośrednia} P3 (lub jej odfiltrowany podzbiór), z tym ograniczeniem,
że oczekuje istnienia w nich zmiennych \code{id_abs} i \code{rok_abs}, kombinację
wartości których traktuje jako identyfikator absolwenta (jednostki).
}
\examples{
\dontrun{
  przygotuj_dane_przeplywy(p3[p3$rok_abs == 2019, ], status, c(6, 30,54))
}
}
