% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splaszcz_wskazniki_pd.R
\name{splaszcz_wskazniki_pd}
\alias{splaszcz_wskazniki_pd}
\title{Przeksztalcanie wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
splaszcz_wskazniki_pd(
  x,
  dopiszAtrybuty = c("nie", "lAbs", "lSzk", "lZadenZWymienionych", "lNieDotyczy")
)
}
\arguments{
\item{x}{ramka danych zawierająca (między innymi) kolumny ze zagregowanymi
wskaźnikami przeznaczonymi do publicznej prezentacji, typowo zwrócona przez
funkcje \code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}} lub
\code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}}}

\item{dopiszAtrybuty}{wektor ciągów znaków podający nazwy atrybutów
wskaźników, które mają zostać dodane jako dodatkowe kolumny do tworzonej
\emph{płaskiej} reprezentacji lub "nie" aby wskazać, że takie kolumny nie mają
być tworzone}
}
\value{
Ramka danych przekazana argumentem \code{x}, w której wszystkie wskaźniki
zostały \emph{rozwinięte} w kolumny.
}
\description{
Funkcja pozwala \emph{spłaszczyć} strukturę, w jakiej przechowywane są wskaźniki
zagregowane do postaci (potencjalnie bardzo) szerokiej ramki danych.
}
\details{
Funkcja przekształca \strong{wszystko, co \emph{wyglądają jej na zagregowany wskaźnik
przeznaczony do publicznej prezentacji}}, co w praktyce oznacza, że
sprawdzane są elementy kolumn-list w ramce danych przekazanej argumentem \code{x}.

Jeśli argument \code{dopiszZadenZWymienionych=TRUE}, przed dokonaniem opisanego
wyżej przekształcenia, wartość atrybutu \code{lZadenZWymienionych}
(o ile istnieje) zostanie dopisana jako dodatkowy element na końcu wskaźnika.
}
