% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_wskazniki_pd_jst.R
\name{uruchom_oblicz_wskazniki_pd_grupy}
\alias{uruchom_oblicz_wskazniki_pd_grupy}
\title{Nieeksportowane funkcje wykorzystywane w agregowaniu wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
uruchom_oblicz_wskazniki_pd_grupy(
  wierszMatrycy,
  p4,
  p3,
  zmGrupujace,
  zmBezOgolem,
  zmTylkoWartosciWDanych,
  zmWskaznikiP4,
  zmWskaznikiP3,
  statystyki,
  wskTylkoNiezerowe,
  etykietaBrakDanych,
  etykietaOgolem,
  wyswietlPostep,
  liczbaWatkow
)
}
\arguments{
\item{wierszMatrycy}{ramka danych o jednym wierszu, zawierająca kolumny
opisujące zmienne opisujące podział na JST oraz kolumnę-listę \code{wskazniki},
w której zostaną umieszczone wyniki wywołania
\code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}}.}

\item{p4}{ramka danych z tabelą \emph{pośrednią} P4 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs} i \code{id_szk}}

\item{p3}{ramka danych z tabelą \emph{pośrednią} P3 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs}
i \code{mies_od_ukoncz} lub \code{NULL} (jeśli \code{zmWskaznikiP3 } nie podaje żadnych
wskzaźników)}

\item{zmGrupujace}{wektor tekstowy zawierający nazwy zmiennych z ramki danych
przekazanej argumentem \code{p4}, kombinacje wartości których określają grupy,
w ramach których mają zostać obliczone zagregowane wskaźniki (funkcja
poradzi sobie również z przekazaniem tu wektora tekstowego zerowej długości
lub wartości \code{NULL} - poskutkuje to obliczeniem zagregowanych wskaźników na
całości danych - ale takie wywołanie jest oczywiście nietypowe)}

\item{zmBezOgolem}{opcjonalnie wektor tekstowy zawierający nazwy zmiennych,
dla których mają nie być tworzone grupy "Ogółem" opisujące wszystkich
absolwentów, bez względu na konkretną wartość danej zmiennej (podane nazwy
\strong{nie} muszą występować wśród nazw przekazanych argumentem \code{zmGrupujace}
\itemize{
\item jeśli ich tam nie ma, zostaną zignorowane); domyślnie "typ_szk";
przekazanie wektora tekstowego zerowej długości lub wartości \code{NULL} oznacza,
że nie ma być żadnych takich zmiennych
}}

\item{zmTylkoWartosciWDanych}{opcjonalnie lista z nazwanymi elementami,
z których każdy jest wektorem tekstowym, opisujących dla jakich zestawów
zmiennych mają zostać utworzone grupy tylko dla tych kombinacji wartości,
które występują w danych wejściowych przekazanych argumentem \code{p4} (każdy
element definiuje oddzielony zestaw zmiennych) - przy sprawdzaniu, czy dana
kombinacja wartości występuje w danych używane będą zarówno zmienne, których
nazwy zawiera dany element listy, (np. domyślnie \code{kod_zaw} i \code{nazwa_zaw}) jak
zmienna, której nazwę podano nazwą tego elementu listy (np.domyślnie
\code{typ_szk}), przy czym:
\itemize{
\item zmienne podane wartościami elementu listy będą miały tworzoną wspólnie
\emph{wartość ogółem} (tj. wśród tworzonych grup albo wszystkie te zmienne
będą mieć przypisaną \emph{wartość ogółem}, albo żadna z nich; uwaga, to
oznacza, że jeśli choć jedna z nich zostanie też wymieniona
w argumencie \code{zmBezOgolem}, to nie zostaną utworzone żadne grupy,
dla których te zmienne przyjmowały by \emph{wartość ogółem}),
\item dla zmiennej, której nazwa została podana nazwą elementu listy
\emph{wartość ogółem} zostanie utworzona niezależnie od zmiennych, których
nazwy podano elementem listy.
}

Zmienne o nazwach podanych tym argumentem \strong{nie} muszą występować wśród nazw
zmiennych przekazanych argumentem \code{zmGrupujace} - te, z nich, których wśród
nie ma wśród tych drugich, zostaną zignorowane, a kombinacje będą
rozpatrywane ze względu na pozostałe.

Domyślna wartość opisuje, że mają zostać utworzone tylko występujące w danych
kombinacje wartości zmiennych \code{typ_szk}, \code{kod_zaw} i \code{nazwa_zaw}. Jeśli
nie chce się definiować takich zestawów, można przekazać \code{NULL} lub listę
zerowej długości.}

\item{zmWskaznikiP4}{opcjonalnie wektor ciągów znaków z nazwami zmiennych,
które zawierają wskaźniki w danych przekazanych argumentem \code{p4}, które mają
zostać zagregowane; jeśli argument nie zostanie podany, zostanie wykorzystany
domyślny zestaw zmiennych; można podać \code{NULL} lub wektor tekstowy zerowej
długości, aby wskazać, że żadne wskaźniki z \code{p4} nie mają być agregowane}

\item{zmWskaznikiP3}{opcjonalnie wektor ciągów znaków z nazwami zmiennych,
które zawierają wskaźniki w danych przekazanych argumentem \code{p3}, które mają
zostać zagregowane; jeśli argument nie zostanie podany, zostanie wykorzystany
domyślny zestaw zmiennych; można podać \code{NULL} lub wektor tekstowy zerowej
długości, aby wskazać, że żadne wskaźniki z \code{p3} nie mają być agregowane}

\item{statystyki}{opcjonalnie lista z nazwanymi elementami, opisującymi
statystyki, które mają zostać obliczone w odniesieniu do wskaźników
\strong{liczbowych} (w formacie takim, jak argument \code{.fns} w formie listy
w wywołaniach \code{\link[dplyr:across]{dplyr::across()}}); funkcje/wyrażenia mogą zwracać więcej niż
jedną wartość, ale ew. nazwy elementów zwróconego przez funkcję wektora nie
zostaną zachowane; domyślnie średnia obliczana z pominięciem
braków danych}

\item{wskTylkoNiezerowe}{opcjonalnie wektor ciągów znaków z nazwami
\strong{wskaźników} obliczanych w formie rozkładów (a więc na podstawie zmiennych
będących czynnikami, wektorami tekstowymi lub macierzami), dla których mają
zostać zwrócone tylko wartości rozkładu nie będące zerami ani brakami danych}

\item{etykietaBrakDanych}{opcjonalnie ciąg znaków, którym zostaną zastąpione braki
danych w zmiennych grupujących (podanych argumentem \code{zmGrupujace}),
z wyjątkiem tych, które \strong{jednocześnie} są zmiennymi liczbowymi lub
logicznymi i zostały podane argumentem \code{zmBezOgolem}; domyślnie "Ndt."}

\item{etykietaOgolem}{opcjonalnie ciąg znaków, który będzie używany jako
\emph{\strong{wartość ogółem}}, tj. etykieta (wartość) opisująca, że w danym
zestawieniu zostały uwzględnione wszystkie wartości danej zmiennej
grupującej; domyślnie "Ogółem"}

\item{wyswietlPostep}{opcjonalnie wartość logiczna - czy wyświetlać pasek
postępu obrazujący postęp obliczeń? (domyślnie \code{TRUE})
\code{poziom})}

\item{liczbaWatkow}{opcjonalnie dodatnia liczba całkowita - liczba wątków,
w których ma być wywoływana funkcja \code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}}; uwaga, jeśli
dane przekazywane argumentami \code{p4} i \code{p3} są relatywnie \emph{małe} (np.
absolwenci z jednego powiatu), wielowątkowość może wręcz wydłużać czas
obliczeń - p. sekcja \emph{details} w dokumentacji funkcji
\code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}; ogólnie biorąc, niezależnie od warunków wartości
większe niż 3 raczej nie prowadzą już do odczuwalnego skrócenia czasu
obliczeń; domyślnie 1, tj. bez zrównoleglania}
}
\value{
\code{wierszMatrycy} z wypełnioną kolumną \code{wskazniki}.
}
\description{
Dla zadanej kombinacji zmiennych grupujących funkcja odpowiednio zawęża dane
i wywołuje na nich \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}}. Wykorzystywana wewnątrz
\code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}.
}
\details{
Funkcja została wyodrębniona na potrzeby zrównoleglania wywoływania
\code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}} wewnątrz \code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}.
}
