% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usun_puste_wskazniki_pd.R
\name{usun_puste_wskazniki_pd}
\alias{usun_puste_wskazniki_pd}
\title{Przeksztalcanie wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
usun_puste_wskazniki_pd(x, usunZanonimizowane = FALSE)
}
\arguments{
\item{x}{ramka danych zawierająca (między innymi) kolumny ze zagregowanymi
wskaźnikami przeznaczonymi do publicznej prezentacji, typowo zwrócona przez
funkcje \code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}} lub
\code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}}, przy czym ta ostatnia musiała zostać wywołana
z argumentem \code{format="długi"}}

\item{usunZanonimizowane}{wartość logiczna - czy usuwać również wiersze
opisujące kombinacje cech absolwentów, które wystąpiły w danych, ale zostały
zanonimizowane (z wykorzystaniem \code{\link[=zanonimizuj_wskazniki_pd]{zanonimizuj_wskazniki_pd()}}), gdyż liczyły
zbyt mało absolwentów}
}
\value{
Ramka danych przekazana argumentem \code{x}, z której usunięte zostały
wiersze opisujące takie kombinacje cech absolwentów, które nie wysępowały
w danych.
}
\description{
Funkcja pozwala usunąć z zestawienia zagregowanych wskaźników (w formacie
\emph{długim}) wiersze opisujące takie kombinacje cech absolwentów, które nie
wysępowały w danych.
}
