% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapisz_wskazniki_pd_jst.R
\name{zapisz_wskazniki_pd_jst}
\alias{zapisz_wskazniki_pd_jst}
\title{Zapisywanie na dysku wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
zapisz_wskazniki_pd_jst(
  x,
  katalogZapis = "./",
  format = "JSON",
  komunikatCenzura = eval(formals(przygotuj_wskazniki_pd_toJSON)$komunikatCenzura),
  usunKombinacjeBezAbs = TRUE,
  komunikatBrakDanych = eval(formals(przygotuj_wskazniki_pd_toJSON)$komunikatBrakDanych),
  zmTERYT = intersect(c("teryt_woj_szk", "teryt_pow_szk"), names(x)),
  przeliczNaProcenty = TRUE,
  wykluczZprzeliczaniaNaProcenty = c("średnia", "liczba", "suma"),
  pretty = FALSE,
  wyswietlPostep = TRUE
)
}
\arguments{
\item{x}{ramka danych zwrócona przez funkcję \code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}}

\item{katalogZapis}{opcjonalnie ciąg znaków z nazwą katalogu, do którego
mają zostać zapisane pliki; domyślnie bieżący katalog roboczy}

\item{format}{opcjonalnie ciąg znaków wskazujący format zapisu plików -
obecnie jedyna możliwa wartość to "JSON"}

\item{komunikatCenzura}{opcjonalnie ciąg znaków - komunikat, który ma
zostać zapisany, jako element "cenzura" (elementy "w" i "a" nie zostaną wtedy
zapisane), dla tych wskaźników, które zostały \emph{ocenzurowane}/zanonimizowane
(tj. ich atrybut \code{lAbs} jest mniejszy od zera lub ich atrybut \code{lSzk} jest
mniejszy od zera);
podany ciag znaków może zawierać kody specjalne "\{lAbs\}" i "\{lSzk\}", za
które podstawione zostaną odpowiednie wartości progów użytych podczas
anonimizacji (odczytane z atrybutów \code{lAbs} i \code{lSzk});
jeśli taki atrybut nie ma być dopisywany, należy argumentem przekazać wartość
\code{NA_character_}}

\item{usunKombinacjeBezAbs}{wartość logiczna - czy przed zapisem do plików
z danych powinny zostać usunięte wiersze opisujące kombinacje wartości
zmiennych grupujących, dla których nie wystąpił ani jeden abolwent}

\item{komunikatBrakDanych}{opcjonalnie ciąg znaków - komunikat, który ma
zostać zapisany, jako element "brakDanych" (elementy "w" i "a" nie zostaną
wtedy zapisane), dla tych wskaźników, dla których nie było żadnych
absolwentów, na podstawie których można by obliczyć wskaźnik (tj. atrybut
\code{lAbs} równy jest równy zero lub atrybut \code{lSzk} jest równy zero);
jeśli taki atrybut nie ma być dopisywany, należy argumentem przekazać wartość
\code{NA_character_}}

\item{zmTERYT}{opcjonalnie nazwa zmiennej zawierającej identyfikatory JST
(kody TERYT); jeśli zostanie tu podany wektor zerowej długości, funkcja
przyjmie, że przekazane dane to wyniki ogólnopolskie}

\item{przeliczNaProcenty}{opcjonalnie wartość logiczna - czy przeliczać
wartości wskaźników na procenty? (domyślnie \code{TRUE})}

\item{wykluczZprzeliczaniaNaProcenty}{opcjonalnie wektor ciągów znaków
z nazwami elementów wartości wskaźników, które wskazują, że dana wartość
\strong{nie} powinna zostać przeliczona na procenty (domyślnie "średnia",
"liczba" i "suma")}

\item{pretty}{opcjonalnie wartość logiczna - czy zapisywane pliki JSON
mają być sformatowane w sposób ułatwiający ich czytanie przez człowieka,
tzn. z łamaniem do wielu linii i wcięciami? (domyślnie \code{FALSE}, w celu
zmniejszenia wielkości zapisywanych plików)}

\item{wyswietlPostep}{opcjonalnie wartość logiczna - czy wyświetlać pasek
postępu obrazujący postęp zapisu? (domyślnie \code{TRUE})}
}
\value{
Funkcja zwraca \emph{niewidocznie} wektor nazw zapisanych plików.
}
\description{
Funkcja zapisuje na dysku pliki z wartościami wskaźników obliczonych przez
\code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}) - obecnie jedynym odstępnym formatem są pliki
JSON.
}
\details{
Funkcja została napisana tak, aby miała możliwie niewielki ślad pamięciowy,
gdyż zestawienia przygotowane przy pomocy \code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}} mogą
być ekstremalnie duże (jeśli tylko uzyje się nieco szerszego zestawu
zmiennych grupujących).

Wewnątrz wykorzystywana jest funkcje \code{\link[=przygotuj_wskazniki_pd_toJSON]{przygotuj_wskazniki_pd_toJSON()}},
a następnie \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.
}
