% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodaj_wskazniki_prace.R
\name{dodaj_wskazniki_prace}
\alias{dodaj_wskazniki_prace}
\title{Obliczanie wskaznikow wykorzystywanych w raportach}
\usage{
dodaj_wskazniki_prace(p4, p3)
}
\arguments{
\item{p4}{ramka danych z tabelą \emph{pośrednią} P4 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs} i \code{rok_abs});
jeśli nie ma potrzeby dołączania obliczonych wskaźników do tabeli
\emph{pośredniej} P4, można ustawić wartość tego argumentu na \code{NULL}}

\item{p3}{ramka danych z tabelą \emph{pośrednią} P3 (lub jej odfiltrowanym
podzbiorem)}
}
\value{
Ramka danych przekazana argumentem \code{p4} (lub - jeśli argument \code{p4}
został ustawiony na \code{NULL} - ramka danych z unikalnymi kombinacjami wartości
zmiennych (\code{id_abs}, \code{rok_abs}) występującym w ramce danych przekazanej
argumentem \code{p3}) z dodanymi kolumnami:
\itemize{
\item \code{sr_wynagr_r0_ivkw} - średnie wynagrodzenia w IV kw. roku ukończenia szkoły,
\item \code{sr_wynagr_r1} - średnie wynagrodzenia w roku następującym po roku ukończenia szkoły,
\item \code{sr_wynagr_r1_ikw} - średnie wynagrodzenia w I kw. roku następującym po roku ukończenia szkoły,
\item \code{sr_wynagr_r2_ikw} - średnie wynagrodzenia w I kw. roku następującego w dwa lata po roku ukończenia szkoły,
\item \code{sr_wynagr_r2} - średnie wynagrodzenia w roku następującym w dwa lata po roku ukończenia szkoły,
\item \code{sr_wynagr_r3} - średnie wynagrodzenia w roku następującym w trzy lata po roku ukończenia szkoły,
\item \code{sr_wynagr_r4} - średnie wynagrodzenia w roku następującym w cztery lata po roku ukończenia szkoły,
\item \code{sr_wynagr_uop_nauka_r0_wrzgru} - średnie relatywne (odniesione do średnich miesięcznych wynagrodzeń w powiecie w danym roku) wynagrodzenia z umów o pracę od września do grudnia roku ukończenia szkoły w miesiącach kontynuowania nauki,
\item \code{sr_wynagr_uop_bez_nauki_r0_wrzgru} - średnie relatywne (odniesione do średnich miesięcznych wynagrodzeń w powiecie w danym roku) wynagrodzenia z umów o pracę od września do grudnia roku ukończenia szkoły w miesiącach niekontynuowania nauki,
\item \code{praca_nauka_r0_wrzgru} - długość zatrudnienia w okresie kontynuowania nauki między wrześniem a grudniem roku ukończenia szkoły,
\item \code{praca_bez_nauki_r0_wrzgru} - długość zatrudnienia w okresie bez kontynuowania nauki między wrześniem a grudniem roku ukończenia szkoły,
\item \code{bezrobocie_r0_wrzgru} - długość bycia zarejestrowanym bezrobotnym od września do grudnia roku ukończenia szkoły.
}
}
\description{
Na podstawie tabeli \emph{pośredniej} P3 funkcja oblicza wskaźniki opisujące
średnią wysokość wynagrodzeń oraz długość posiadania pracy lub bycia
bezrobotnym (w wybranych okresach).
}
\details{
W porównaniu do pozostałych funkcji \code{dodaj_[...]()} działanie tej jest
najmniej elastyczne, ze zdefiniowanymi na stałe okresami, dla których
obliczane są poszczególne wskaźniki. Zasadniczo jest ona przeznaczona do
użycia w ramach procesu przygotowywania tabel \emph{pośrednich} przez
\code{\link[MLASdaneAdm:przygotuj_tabele_posrednie]{MLASdaneAdm::przygotuj_tabele_posrednie()}} i późniejsze jej użycie jest
bezcelowe.
}
\seealso{
\code{\link[=oblicz_wskaznik_z_p3]{oblicz_wskaznik_z_p3()}} - \emph{koń roboczy} odpowiedzialny za agregację
wskaźników w ramach \code{dodaj_wskazniki_prace()}
}
