% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legalneKontynuacje}
\alias{legalneKontynuacje}
\title{Legalne kontynuacje nauki}
\format{
Ramka danych z 26 wierszami i 2 kolumnami:
\itemize{
\item{\code{typ_szk} - typ ukończonej szkoły (czynnik),}
\item{\code{typ_szk_kont} - forma kontynuacji nauki (wektor ciągów znaków).}
}
}
\usage{
legalneKontynuacje
}
\description{
Zestawienie nominalnie jedynych możliwych form kontynuowania
nauki w zależności od typu ukończonej szkoły ponadpodstawowej.
}
\seealso{
\code{\link[=dodaj_wskazniki_kontynuacje]{dodaj_wskazniki_kontynuacje()}}
}
\keyword{datasets}
