% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_wskazniki_pd.R
\name{oblicz_wskazniki_pd}
\alias{oblicz_wskazniki_pd}
\title{Agregowanie wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
oblicz_wskazniki_pd(
  p4,
  p3,
  statystyki = list(średnia = ~mean(., na.rm = TRUE)),
  oddzielCzasOdWskaznika = TRUE,
  format = c("długi", "szeroki"),
  miesiace = NULL
)
}
\arguments{
\item{p4}{ramka danych z tabelą \emph{pośrednią} P4 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs} i \code{id_szk}}

\item{p3}{ramka danych z tabelą \emph{pośrednią} P3 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs}
i \code{mies_od_ukoncz}}

\item{statystyki}{opcjonalnie lista z nazwanymi elementami, opisującymi
statystyki, które mają zostać obliczone w odniesieniu do wskaźników
\strong{liczbowych} (w formacie takim, jak argument \code{.fns} w formie listy
w wywołaniach \code{\link[dplyr:across]{dplyr::across()}}); funkcje/wyrażenia mogą zwracać więcej niż
jedną wartość, ale ew. nazwy elementów zwróconego przez funkcję wektora nie
zostaną zachowane; domyślnie średnia obliczana z pominięciem
braków danych}

\item{oddzielCzasOdWskaznika}{opcjonalnie wartość logiczna - czy w zwracanej
ramce danych nazwa wskaźnika powinna zostać rozbita na dwie kolumny:
\code{wskaznik} i \code{czas} (\code{TRUE}), czy też tylko kolumnę \code{wskaznik}, zawierającą
nazwę wskaźnika połączoną z kodem okresu, dla jakiego został on obliczony
(\code{FALSE})}

\item{format}{opcjonalnie ciąg znaków opisujący, czy wyniki mają zostać
zwrócone w formie \emph{długiej} (domyślnie), czy \emph{szerokiej}}

\item{miesiace}{opcjonalnie wektor liczb całkowitych (co do zasady dodatnich)
opisujący, wiersze o jakich wartościach kolumny \code{mies_od_ukoncz} z ramki
danych przekazanej argumentem \code{p3} mają zostać wykorzystane do obliczenia
zagregowanych wskaźników (wszystkie inne zostaną pominięte); pozwala też
uzyskać efekt polegający na utworzeniu pożądanego zestawy wskaźników
obliczonych na podstawie danych z tabeli \emph{pośredniej} P3 w sytuacji, gdy
dane przekazane do funkcji są puste (mają zero wierszy); domyślnie zostaną
wykorzystane wszystkie dane z \code{p3}}
}
\value{
Ramka danych z obliczonymi wskaźnikami, w zależności od kombinacji
wartości argumentów \code{oddzielCzasOdWskaznika} i \code{format} zawierająca:
\itemize{
\item wiele wierszy i trzy kolumny: \code{wskaznik}, \code{czas} i \code{wartosc}
(domyślnie: \code{oddzielCzasOdWskaznika=TRUE} i \code{format="długi"}),
\item wiele wierszy i dwie kolumny: \code{wskaznik} i \code{wartosc}
(\code{oddzielCzasOdWskaznika=FALSE} i \code{format="długi"}),
\item wiele wierszy, kolumnę \code{czas} i kolejne kolumny zawierające poszczególne
wskaźniki (\code{oddzielCzasOdWskaznika=TRUE} i \code{format="szeroki"}),
\item jeden wiersz i kolejne kolumny zawierające poszczególne
wskażniki (\code{oddzielCzasOdWskaznika=FALSE} i \code{format="szeroki"}).
}

Odnośnie do formatu kolumn zawierających wartości wskaźników, patrz niżej.
}
\description{
Funkcja pozwala obliczyć zestawienia zagregowanych wskaźników z danej edycji
monitoringu dla przekazanych danych. Funkcja agreguje \strong{wszystkie kolumny
w danych, które \emph{wyglądają jej na zawierające wskaźniki}}, co w praktyce
oznacza:
\itemize{
\item dla ramki danych przekazanej argumentem \code{p4}, że ich nazwa albo kończy
się liczbą, albo jest postaci "\emph{r[liczba]}[litery]";
\item dla ramki danych przekazanej argumentem \code{p3}, wszystkie kolumny poza
\code{id_abs}, \code{rok_abs}, \code{mies_od_ukoncz}, \code{rok}, \code{miesiac} i \code{okres};
\strong{dane z poszczególnych miesięcy (identyfikowanych wartościami kolumny
\code{mies_od_ukoncz}) będą tworzyć oddzielne wskaźniki zagregowane} (por.
opis argumentu \code{miesiace}).
}
}
\details{
Agregacja przeprowadzana jest w różny sposób w zależności od typu kolumny
w wejściowych danych:
\itemize{
\item Kolumny będące czynnikami lub wektorami tekstowymi (w sensie sprawdzania
typu z wykorzystaniem funkcji \code{\link[=is.factor]{is.factor()}} i \code{\link[=is.character]{is.character()}}) zostaną
zagregowane z wykorzystaniem funkcji \code{\link[=table]{table()}},
\item Kolumny będące wektorami liczbowymi lub logicznymi (w sensie sprawdzania
typu z wykorzystaniem funkcji \code{\link[=is.numeric]{is.numeric()}} i \code{\link[=is.logical]{is.logical()}}) zostaną
zagregowane przy pomocy funkcji/wyrażeń przekazanych argumentem
\code{statystyki} - efektem będzie wektor, nazwy elementów którego
odpowiadają nazwom elementów listy przekazanej argumentem \code{statystyki}
(domyślnie jednoelementowy wektor, z nazwą elementu "średnia"),
\item Kolumny będące macierzami liczb lub wartości logicznych zostaną
zagregowane poprzez zsumowanie wartości w kolumnach, przy czym
\strong{kolumny, w których znajdowały się tylko braki danych, zostaną usunięte
z wyników} (nie dotyczy to sytuacji, gdy dane są puste, tzn. mają zero
wierszy, bo wtedy nie ma sensu rozważanie, czy w danej kolumnie znajduje
się brak danych, czy nie).
}

W efekcie każda kolumna zawierająca zagregowane wskażniki w zwracanej ramce
danych jest listą. Każdy z jej elementów stanowi wskaźnik obliczony zgodnie
z powyższym opisem, który ma dodatkowo przypisane atrybuty
(p. \code{\link[=attributes]{attributes()}}), z których każdy jest liczbą naturalną:
\itemize{
\item \code{lAbs} - liczba wierszy (absolwentów), na podstawie których został
obliczony zagregowany wskaźnik,
\item \code{lSzk} - liczba różnych szkół, które ukończyli absolwenci, na podstawie
których został obliczony zagregowany wskaźnik,
\item \code{lNieDotyczy} - liczba wierszy (absolwentów), zawierających braki danych
(w przypadku kolumn wejściowych będących macierzami - we wszystkich
kolumnach macierzy),
\item Wskaźniki zagregowane obliczone na podstawie takich kolumn
mają dodatkowy atrybut \code{lZadenZWymienionych}, który przechowuje
liczbę wierszy, w których wszystkie kolumny macierzy były albo
zerami albo brakami danych (ale z pominięciem wierszy zawierających
same braki danych).
}
}
\seealso{
\code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}},
\code{\link[=zanonimizuj_wskazniki_pd]{zanonimizuj_wskazniki_pd()}}
}
