% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_wskazniki_pd_grupy.R
\name{oblicz_wskazniki_pd_grupy}
\alias{oblicz_wskazniki_pd_grupy}
\title{Agregowanie wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
oblicz_wskazniki_pd_grupy(
  p4,
  p3,
  zmGrupujace,
  zmBezOgolem = "typ_szk",
  zmTylkoWartosciWdanych = list(c("typ_szk", "kod_zaw", "nazwa_zaw")),
  statystyki = list(średnia = ~mean(., na.rm = TRUE)),
  wyswietlPostep = TRUE
)
}
\arguments{
\item{p4}{ramka danych z tabelą \emph{pośrednią} P4 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs} i \code{id_szk}}

\item{p3}{ramka danych z tabelą \emph{pośrednią} P3 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs}
i \code{mies_od_ukoncz}}

\item{zmGrupujace}{wektor tekstowy zawierający nazwy zmiennych z ramki danych
przekazanej argumentem \code{p4}, kombinacje wartości których określają grupy,
w ramach których mają zostać obliczone zagregowane wskaźniki (funkcja
poradzi sobie również z przekazaniem tu wektora tekstowego zerowej długości
lub wartości \code{NULL} - poskutkuje to obliczeniem zagregowanych wskaźników na
całości danych - ale takie wywołanie jest oczywiście nietypowe)}

\item{zmBezOgolem}{opcjonalnie wektor tekstowy zawierający nazwy zmiennych,
dla których mają nie być tworzone grupy "Ogółem" opisujące wszystkich
absolwentów, bez względu na konkretną wartość danej zmiennej (podane nazwy
\strong{nie} muszą występować wśród nazw przekazanych argumentem \code{zmGrupujace}
\itemize{
\item jeśli ich tam nie ma, zostaną zignorowane); domyślnie "typ_szk";
przekazanie wektora tekstowego zerowej długości lub wartości \code{NULL} oznacza,
że nie ma być żadnych takich zmiennych
}}

\item{zmTylkoWartosciWdanych}{opcjonalnie lista wektorów tekstowych,
pisujących dla jakich zestawów zmiennych mają zostać utworzone grupy tylko
dla tych kombinacji wartości, które występują w danych wejściowych
przekazanych argumentem \code{p4} - każdy element definiuje oddzielony zestaw
zmiennych (podane nazwy \strong{nie} muszą występować wśród nazw przekazanych
argumentem \code{zmGrupujace} - te, których wśród nich nie ma, zostaną
zignorowane, a kombinacje będą rozpatrywane ze względu na pozostałe);
domyślna wartość opisuje, że mają zostać utworzone tylko występujące w danych
kombinacje wartości zmiennych \code{typ_szk}, \code{kod_zaw} i \code{nazwa_zaw}; jeśli
nie chce się definiować takich zestawów, można przekazać \code{NULL} lub listę
zerowej długości}

\item{statystyki}{opcjonalnie lista z nazwanymi elementami, opisującymi
statystyki, które mają zostać obliczone w odniesieniu do wskaźników
\strong{liczbowych} (w formacie takim, jak argument \code{.fns} w formie listy
w wywołaniach \code{\link[dplyr:across]{dplyr::across()}}); funkcje/wyrażenia mogą zwracać więcej niż
jedną wartość, ale ew. nazwy elementów zwróconego przez funkcję wektora nie
zostaną zachowane; domyślnie średnia obliczana z pominięciem
braków danych}

\item{wyswietlPostep}{opcjonalnie wartość logiczna - czy wyświetlać pasek
postępu obrazujący postęp obliczeń? (domyślnie \code{TRUE})
\code{poziom})}
}
\value{
Ramka danych z kolumnami:
\itemize{
\item \code{rok_abs},
\item kolumnami podanymi argumentem \code{zmGrupujace}, przekształconymi
w czynniki i - z wyjątkiem kolumn podanych argumentem \code{zmBezOgolem} -
z dodaną dodatkową wartością "Ogółem",
\item \code{wskaznik} - kolumna tekstowa opisująca wskaźnik,
\item \code{czas} - kolumna tekstowa opisująca czas, jaki opisuje wskaźnik, co do
zasady miesiąc od ukończenia szkoły ponadpodstawowej (jeśli jest to
liczba) lub skrótowiec opisujący kilka miesięcy (np. "r0_ivkw" -
IV kwartał roku ukończenia szkoły ponadpodstawowej),
\item \code{wartosc} - kolumna-lista zawierająca wartości wskaźników
(p. \code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}})
}
}
\description{
Funkcja pozwala obliczyć zestawienia zagregowanych wskaźników z danej edycji
monitoringu w ramach każdej grupy zdefiniowanej przez kombinację wartości
podanych zmiennych (\emph{grupujących/niezależnych}).
}
\seealso{
\code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}},
\code{\link[=zanonimizuj_wskazniki_pd]{zanonimizuj_wskazniki_pd()}}
}
