% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_wynagrodzenia_wzgledne.R
\name{oblicz_wynagrodzenia_wzgledne}
\alias{oblicz_wynagrodzenia_wzgledne}
\title{Obliczanie wzglednych wskaznikow wynagrodzen}
\usage{
oblicz_wynagrodzenia_wzgledne(
  p3,
  miesOdUkoncz,
  nazwaWsk = "sr_wynagr_uop",
  nauka2 = NA
)
}
\arguments{
\item{p3}{ramka danych z tabelą \emph{pośrednią} P3 (lub jej odfiltrowanym
podzbiorem)}

\item{miesOdUkoncz}{wektor liczbowy podający wartości zmiennej
\code{miesOdUkoncz}, które mają być uwzględnione przy obliczaniu wartości
wskaźnika}

\item{nazwaWsk}{opcjonalnie ciąg znaków z nazwą kolumny, w której mają się
znaleźć obliczone wartości wskaźnika względynych wynagrodzeń (domyślnie
"sr_wynagr_uop")}

\item{nauka2}{opcjonalnie pojedyncza wartość liczbowa opisująca, czy dla
danego absolwenta mają być uwzględnione tylko miesiące, w których kontynuował
on naukę (\code{nauka2 = 1}) albo tylko miesiące, w których nie kontynuował on
nauki (\code{nauka2 = 0}); domyślna wartość (\code{NA}) wskazuje, że mają zostać
uwzględnione wszystkie miesiące}
}
\value{
Ramka danych o kolumnach takich jak w \code{p3} z dodaną kolumną o nazwie
podanej argumentem \code{nazwaWsk}. Liczba wierszy odpowiada liczbie unikalnych
kombinacji wartości zmiennych (\code{id_abs}, \code{rok_abs}) w obiekcie przekazanym
argumentem \code{p3}.
}
\description{
Funkcja oblicza wartości wskaźnika względnego wynagrodzeń za podany okres
(i ew. biorąc pod uwagę wyłącznie miesiące nauki albo bez nauki) na poziomie
poszczególnych absolwentów, na podstawie danych z tabeli \emph{pośredniej} P3.
}
\details{
Względne wynagrodzenia obliczane są jako iloraz
\code{wynagrodzenie_uop / powiat_sr_wynagrodzenie}, a następnie przekazywane do
funkcji \code{\link[=oblicz_wskaznik_z_p3]{oblicz_wskaznik_z_p3()}} w celu dokonania agregacji.

Ocena, czy absolwent kontynuował naukę dokonywana jest na podstawie zmiennej
\code{nauka2}, co oznacza, że uczestnictwo w KKZ lub KUZ \strong{nie} jest traktowane
jako kontynuacja nauki (w przypadku KKZ - o ile nie wiąże się z nauką w BS II).

Funkcja została wyodrębiona w dużej mierze po to, aby mogła być użyta
również w ramach pakietu \emph{MLASdaneAdm}, wewnątrz funkcji
\code{\link[MLASdaneAdm:uzupelnij_wzgledne_wynagrodzenia]{MLASdaneAdm::uzupelnij_wzgledne_wynagrodzenia()}}.
}
