% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_wskazniki_pd_toJSON.R
\name{przygotuj_wskazniki_pd_toJSON}
\alias{przygotuj_wskazniki_pd_toJSON}
\title{Agregowanie wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
przygotuj_wskazniki_pd_toJSON(x)
}
\arguments{
\item{x}{ramka danych zawierająca (między innymi) kolumny ze zagregowanymi
wskaźnikami przeznaczonymi do publicznej prezentacji, typowo zwrócona przez
funkcje \code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}} lub
\code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}}}
}
\value{
Ramka danych przekazana argumentem \code{x}, w której wskaźniki zostały
przekształcone na zagnieżdżone listy.
}
\description{
Funkcja przekształca wskaźniki obliczone przez funkcję \code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}}
(a więc również przez \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}} lub
\code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}) do formatu, który będzie przyjazny zapisaniu
ich w formacie JSON przy pomocy funkcji \code{toJSON()} z pakietu \emph{jsonlite}.
}
