% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_wskazniki_pd_grupy.R
\name{uruchom_oblicz_wskazniki_pd}
\alias{uruchom_oblicz_wskazniki_pd}
\title{Nieeksportowane funkcje wykorzystywane w agregowaniu wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
uruchom_oblicz_wskazniki_pd(
  wierszMatrycy,
  p4,
  p3,
  etykietaOgolem,
  zmWskaznikiP4,
  zmWskaznikiP3,
  statystyki,
  wskTylkoNiezerowe,
  miesiace
)
}
\arguments{
\item{wierszMatrycy}{ramka danych o jednym wierszu, zawierająca kolumny
opisujące zmienne grupujące oraz kolumnę-listę \verb{___wskazniki___}, w której
zostaną umieszczone wyniki wywołania \code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}}.}

\item{p4}{ramka danych z tabelą \emph{pośrednią} P4 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs} i \code{id_szk}}

\item{p3}{ramka danych z tabelą \emph{pośrednią} P3 (lub jej odfiltrowanym
podzbiorem), musi zawierać co najmniej kolumny \code{id_abs}, \code{rok_abs}
i \code{mies_od_ukoncz}}

\item{etykietaOgolem}{opcjonalnie ciąg znaków, który będzie używany jako
\emph{\strong{wartość ogółem}}, tj. etykieta (wartość) opisująca, że w danym
zestawieniu zostały uwzględnione wszystkie wartości danej zmiennej
grupującej; domyślnie "Ogółem"}

\item{zmWskaznikiP4}{opcjonalnie wektor ciągów znaków z nazwami zmiennych,
które zawierają wskaźniki w danych przekazanych argumentem \code{p4}, które mają
zostać zagregowane; jeśli argument nie zostanie podany, zostanie wykorzystany
domyślny zestaw zmiennych; można podać \code{NULL} lub wektor tekstowy zerowej
długości, aby wskazać, że żadne wskaźniki z \code{p4} nie mają być agregowane}

\item{zmWskaznikiP3}{opcjonalnie wektor ciągów znaków z nazwami zmiennych,
które zawierają wskaźniki w danych przekazanych argumentem \code{p3}, które mają
zostać zagregowane; jeśli argument nie zostanie podany, zostanie wykorzystany
domyślny zestaw zmiennych; można podać \code{NULL} lub wektor tekstowy zerowej
długości, aby wskazać, że żadne wskaźniki z \code{p3} nie mają być agregowane}

\item{statystyki}{opcjonalnie lista z nazwanymi elementami, opisującymi
statystyki, które mają zostać obliczone w odniesieniu do wskaźników
\strong{liczbowych} (w formacie takim, jak argument \code{.fns} w formie listy
w wywołaniach \code{\link[dplyr:across]{dplyr::across()}}); funkcje/wyrażenia mogą zwracać więcej niż
jedną wartość, ale ew. nazwy elementów zwróconego przez funkcję wektora nie
zostaną zachowane; domyślnie średnia obliczana z pominięciem
braków danych}

\item{wskTylkoNiezerowe}{opcjonalnie wektor ciągów znaków z nazwami
\strong{wskaźników} obliczanych w formie rozkładów (a więc na podstawie zmiennych
będących czynnikami, wektorami tekstowymi lub macierzami), dla których mają
zostać zwrócone tylko wartości rozkładu nie będące zerami ani brakami danych}

\item{miesiace}{opcjonalnie wektor liczb całkowitych (co do zasady dodatnich)
opisujący, wiersze o jakich wartościach kolumny \code{mies_od_ukoncz} z ramki
danych przekazanej argumentem \code{p3} mają zostać wykorzystane do obliczenia
zagregowanych wskaźników (wszystkie inne zostaną pominięte); pozwala też
uzyskać efekt polegający na utworzeniu pożądanego zestawy wskaźników
obliczonych na podstawie danych z tabeli \emph{pośredniej} P3 w sytuacji, gdy
dane przekazane do funkcji są puste (mają zero wierszy); domyślnie zostaną
wykorzystane wszystkie dane z \code{p3}}
}
\value{
\code{wierszMatrycy} z wypełnioną kolumną \verb{___wskazniki___}.
}
\description{
Dla zadanej kombinacji zmiennych grupujących funkcja odpowiednio zawęża dane
i wywołuje na nich \code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}}. Wykorzystywana wewnątrz
\code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}}.
}
\details{
Funkcja została wyodrębniona na potrzeby zrównoleglania wywoływania
\code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}} wewnątrz \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}}.
}
