% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zanonimizuj_wskazniki_pd.R
\name{zanonimizuj_wskazniki_pd}
\alias{zanonimizuj_wskazniki_pd}
\title{Anonimizowanie wskaznikow w nierozlacznych podgrupach do publicznej prezentacji}
\usage{
zanonimizuj_wskazniki_pd(x, progAbs = 10, progSzk = 3)
}
\arguments{
\item{x}{ramka danych zawierająca (między innymi) kolumny ze zagregowanymi
wskaźnikami przeznaczonymi do publicznej prezentacji, typowo zwrócona przez
funkcje \code{\link[=oblicz_wskazniki_pd_jst]{oblicz_wskazniki_pd_jst()}}, \code{\link[=oblicz_wskazniki_pd_grupy]{oblicz_wskazniki_pd_grupy()}} lub
\code{\link[=oblicz_wskazniki_pd]{oblicz_wskazniki_pd()}}}

\item{progAbs}{liczba dodatnia - liczba absolwentów poniżej której wskaźniki
zostaną zanonimizowane}

\item{progSzk}{liczba dodatnia - liczba szkół, poniżej której wskaźniki
zostaną zanonimizowane}
}
\value{
Ramka danych przekazana argumentem \code{x}, w której wskaźniki zostały
zanonimizowane.
}
\description{
Funkcja pozwala zanonimizować wskaźniki zagregowane (zamienić ich wartości na
braki danych) przeznaczone do publicznej prezentacji, które zostały obliczone
na podstawie danych opisujących mniej niż zadaną liczbę absolwentów lub
na podstawie absolwentów mniej niż zadanej liczby różnych szkół.

Warto zwrócić uwagę, że taka anonimizajca ignoruje ewentualną możliwość
określenia wartości wskaźnika poprzez znajomość wartości wskaźników w grupie
ogólniejszej i we wszystkich innych grupach poza daną (implementacja
anonimizacji uwzględniającej ten fakt była by jednak dużo bardziej
skomplikowana).
}
\details{
Funkcja anonimizuje \strong{wszystko, co \emph{wyglądają jej na zagregowany wskaźnik
przeznaczony do publicznej prezentacji}}, co w praktyce oznacza, że
sprawdzane są elementy kolumn-list w ramce danych przekazanej argumentem \code{x}.
Jeśli element takiej kolumny listy ma choć jeden z atrybutów nazywających się
\code{lAbs} lub \code{lSzk}, ich wartości zostaną przyrównane do progów podanych
argumentami odpowiednio \code{probAbs} i \code{progSzk} w wywołaniu funkcji. Jeśli
wartość choć jednego z tych atrybutów jest mniejsza od odpowiedniej wartości
progu (\strong{ale większa od zera}), to wartości danego elementu listy zostaną
zamienione na braki danych (jeśli jest to wektor czy \emph{array}, wszystkie jego
elementy zostaną zamienione na braki danych), tak samo jak wartości atrybutów
\code{lZadenZWymienionych} i \code{lNieDotyczy} (jeśli były zdefiniowane); z kolei
wartości atrybutów \code{lAbs} i \code{lSzk} zostaną zamienione na wartości -1 razy
odpowiednio \code{probAbs} i \code{progSzk}.

\strong{Wskaźniki obliczone na podstawie pustych grup absolwentów nie są
anonimizowane!}
}
